// api/models/wallet.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PardoMart Node API
 * API documentation for the PardoMart Node.js application.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@pardomart.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface Wallet {
    'id'?: string;
    'userId'?: string;
    'balance'?: number;
    'createdAt'?: string;
    'updatedAt'?: string;
}