// api/models/wallet-transaction.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PardoMart Node API
 * API documentation for the PardoMart Node.js application.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@pardomart.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { TransactionStatus } from './transaction-status';
// May contain unused imports in some cases
// @ts-ignore
import type { TransactionType } from './transaction-type';

export interface WalletTransaction {
    'id'?: string;
    'walletId'?: string;
    /**
     * Positive for credit, negative for debit.
     */
    'amount'?: number;
    'type'?: TransactionType;
    'status'?: TransactionStatus;
    'description'?: string | null;
    /**
     * Extra metadata, like an order ID.
     */
    'meta'?: object | null;
    'createdAt'?: string;
    'updatedAt'?: string;
}